requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.7
flask-wtf==1.2.1
flask-cors==4.0.1
python-dotenv==1.0.1
bcrypt==4.1.3
wtforms==3.1.2
reportlab==4.2.2
qrcode==7.4.2
requests==2.32.3
pillow==10.4.0
gunicorn==22.0.0
